[Net] Move CAddrDB out of net.h/.cpp
In accordance to BitCoin Commit d93b14dc5ddfb937b0cc18be425b9d048cefb66b
