+++ src/Makefile.am
@@ -80,6 +80,7 @@ endif
 .PHONY: FORCE
 # Silk Core #
 SILK_CORE_H = \
+  addrdb.h \
   addrman.h \
   alert.h \
   allocators.h \
@@ -170,6 +171,7 @@ libsilk_util_a-clientversion.$(OBJEXT): obj/build.h
 libsilk_server_a_CPPFLAGS = $(SILK_INCLUDES) $(MINIUPNPC_CPPFLAGS)
 libsilk_server_a_SOURCES = \
   addrman.cpp \
+  addrdb.cpp \
   alert.cpp \
   bloom.cpp \
   chain.cpp \
